Add dft function

+++ src/programs/filter/Clipping.jl
@@ -27,8 +27,8 @@ Clipping
 """
 function clipping(width::Int, height::Int, pixels::Matrix{Float64})::Matrix{Int16}
     img::Matrix{Int16} = round.(Int16, pixels)
-    for i in 1:height
-        for j in 1:width
+    for i::Int in 1:height
+        for j::Int in 1:width
             if img[i, j] < 0
                 img[i, j] = 0
             elseif img[i, j] > 255
@@ -59,8 +59,8 @@ Clipping
 - `img::Matrix{Int16}`
 """
 function clipping(width::Int, height::Int, pixels::Matrix{Int16})::Matrix{Int16}
-    for i in 1:height
-        for j in 1:width
+    for i::Int in 1:height
+        for j::Int in 1:width
             if pixels[i, j] < 0
                 pixels[i, j] = 0
             elseif pixels[i, j] > 255
